# docs/swagger.yaml
basePath: /api
definitions:
...
  helpers.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.Brand:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        type: string
      deleted_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      modified_at:
        type: string
      modified_by:
        type: integer
      name:
        type: string
    type: object
  models.BrandCreateDTO:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
  models.BrandUpdateDTO:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
info:
  contact:
    email: developer.haritsnb@gmail.com
    name: Harits Nala Barrun
  description: Backend RESTful API Sistem Kasir & Manajemen Rental Alat Camping menggunakan
    Golang + Gin + PostgreSQL + JWT Blacklist + KYC Upload.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Outdoor Gear Rental POS API
  version: "1.0"
paths:
...
  /brands:
    get:
      description: Memuat seluruh brand peralatan outdoor yang terdaftar
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Brand'
                  type: array
              type: object
        "401":
          description: Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Daftar Semua Brands
      tags:
      - Brands
    post:
      consumes:
      - application/json
      description: Menambahkan brand outdoor baru ke katalog (Admin/Staff Only)
      parameters:
      - description: Payload Brand
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BrandCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Brand'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Tambah Brand Baru
      tags:
      - Brands
  /brands/{id}:
    delete:
      description: Menonaktifkan data brand (Admin Only)
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Soft Delete Brand
      tags:
      - Brands
    get:
      description: Memuat rincian informasi brand berdasarkan ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Brand'
              type: object
        "401":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Detail Brand Berdasarkan ID
      tags:
      - Brands
    post:
      description: Memulihkan kembali brand yang berstatus soft-deleted (Admin Only)
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Restore Brand
      tags:
      - Brands
    put:
      consumes:
      - application/json
      description: Memperbarui nama, deskripsi, atau status aktif brand (Admin/Staff
        Only)
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload Update Brand
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BrandUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Brand (Flexible Partial)
      tags:
      - Brands
  /brands/{id}/force:
    delete:
      description: Menghapus fisik data brand secara permanen dari database (Admin
        Only)
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request / Foreign Key Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request / Foreign Key Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request / Foreign Key Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Permanent Brand (Force Delete)
      tags:
      - Brands
  /brands/deleted:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Brand'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Brand Terhapus (Trash)
      tags:
      - Brands
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token_jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"